move resources to qt folder
