+++ napari/_qt/qt_viewer.py
@@ -10,7 +10,6 @@ from qtpy.QtWidgets import QFileDialog, QSplitter, QVBoxLayout, QWidget
 
 from ..components.camera import Camera
 from ..components.layerlist import LayerList
-from ..resources import get_stylesheet
 from ..utils import config, perf
 from ..utils.interactions import (
     ReadOnlyWrapper,
@@ -24,6 +23,7 @@ from ..utils.key_bindings import components_to_key_combo
 from ..utils.theme import get_theme, template
 from .dialogs.qt_about_key_bindings import QtAboutKeyBindings
 from .dialogs.screenshot_dialog import ScreenshotDialog
+from .resources import get_stylesheet
 from .tracing.qt_performance import QtPerformance
 from .utils import QImg2array, circle_pixmap, square_pixmap
 from .widgets.qt_dims import QtDims
