+++ napari/_qt/resources/styles/00_base.qss
