+++ napari/_qt/resources/styles/01_buttons.qss
