+++ napari/_qt/resources/styles/02_custom.qss
