+++ napari/_qt/resources/test_resources.py
