+++ napari/_vispy/__init__.py
@@ -1,3 +1,17 @@
+import logging
+
+import qtpy
+from vispy import app
+
+# set vispy application to the appropriate qt backend
+app.use_app(qtpy.API_NAME)
+del app
+
+# set vispy logger to show warning and errors only
+vispy_logger = logging.getLogger('vispy')
+vispy_logger.setLevel(logging.WARNING)
+
+
 from .quaternion import quaternion2euler
 from .utils import create_vispy_visual
 from .vispy_axes_visual import VispyAxesVisual
